refactor: split integrator into separate module

# Conflicts:
#	src/galax/potential/_potential/base.py
