+++ src/galax/dynamics/_dynamics/mockstream/mockstream_generator.py
@@ -14,7 +14,7 @@ from jax.lib.xla_bridge import get_backend
 from jaxtyping import Array, Shaped
 
 from galax.dynamics._dynamics.orbit import Orbit
-from galax.integrate._base import Integrator
+from galax.integrate._api import Integrator
 from galax.integrate._builtin import DiffraxIntegrator
 from galax.potential._potential.base import AbstractPotentialBase
 from galax.typing import BatchVec6, FloatScalar, IntScalar, Vec6, VecN, VecTime
