+++ src/galax/integrate/__init__.py
@@ -1,11 +1,13 @@
 """galax: Galactic Dynamix in Jax."""
 
 
-from galax.integrate import _base, _builtin
+from galax.integrate import _api, _base, _builtin
+from galax.integrate._api import *
 from galax.integrate._base import *
 from galax.integrate._builtin import *
 
 __all__: list[str] = []
+__all__ += _api.__all__
 __all__ += _base.__all__
 __all__ += _builtin.__all__
